pyproject.toml
[project]
name = "khc-cli"
version = "0.1.0"
description = "CLI pour l'analyse et la curation de listes awesome pour les technologies durables et open source"
readme = "README.md"
authors = [
    { name = "Abderrahim Guennoune", email = "aguennoune@outlook.com" }
]
requires-python = ">=3.10.13"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", 
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]>=0.9.0,!=0.16.0",
    "rich>=13.6.0",
    "pygithub>=2.1.1",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "termcolor>=2.3.0",
    "typing-extensions>=4.8.0",
    "markdown>=3.8",
    "twine>=6.1.0",
]

[project.urls]
"Homepage" = "https://github.com/Krypto-Hashers-Community/khc-cli"
"Bug Tracker" = "https://github.com/Krypto-Hashers-Community/khc-cli/issues"

[project.scripts]
khc-cli = "khc_cli.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]
ignore = []
